Updating the index file
