--- Xonnect-Weather-App/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,20 +1,0 @@
-#### What does this PR do?
-
-
-#### Description of Task to be completed?
-
-
-#### How should this be manually tested?
-
-
-#### Any background context you want to provide?
-
-
-#### What are the relevant pivotal tracker stories?
-
-
-#### Screenshots (if appropriate)
-
-
-#### Questions:
-
