--- Xonnect-Weather-App/README.md
@@ -1,2 +1,0 @@
-# Xonnect-Weather-App
-Its shows weather details
