--- Xonnect-Weather-App/client/public/css/style.css
@@ -1,25 +1,0 @@
-body {
-background: url("../img/london.jpg") no-repeat center center fixed;
-background-size: cover;
-}
-.card-color {
-    background: rgba(0, 0, 0, 0.4);
-    border-radius: 7px;
-    width: 500px;
-    margin-top: 130px;
-    padding: 30px;
-    color: white;
-}
-
-.icon-size {
-    font-size: 5rem;
-    position: absolute;
-    margin-left: 22rem;
-}
-
-.mt-input {
-    margin-top: 130px;
-    width: 23rem;
-    border-radius: 5px;
-    padding: 5px;
-}
