--- Xonnect-Weather-App/client/public/index.html
@@ -1,68 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <link rel="stylesheet" href="css/bootstrap.css" />
-    <link rel="stylesheet" href="css/weather-icons.min.css" />
-    <link rel="stylesheet" href="css/style.css" />
-    <title>Weather Application</title>
-</head>
-
-<body>
-    <div class="container-fluid bg-secondary">
-        <div class="container">
-            <nav class="navbar navbar-expand-lg navbar-light">
-                <a class="navbar-brand text-white" href="#">Weather<span class="text-success">Map</span></a>
-                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav"
-                    aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                </button>
-            </nav>
-        </div>
-    </div>
-    <div class="container">
-        <div class="row">
-            <div class="col">
-                <div class="card-color">
-                    <div class="row">
-                        <div class="col-5">
-                            <h4 id="cityname">City Name</h4>
-                            <h2><span id="temp-value">0</span>&deg;</h2>
-                        </div>
-                        <div class="col mt-5">
-                            <h5 id="weather-report">Weather</h5>
-                            <div>Pressure: <span id="pres-val">0</span>hPa</div>
-                            <div>Wind speed: <span id="wind-val">0</span>m/s</div>
-                            <div>Humidity: <span id="hum-val">0</span>%</div>
-                        </div>
-                        <i class="wi wi-day-rain icon-size"></i>
-                    </div>
-                </div>
-
-            </div>
-
-            <div class="col ">
-                <input type="text" id="search-area" placeholder="Type your city here" class="mt-input" required="">
-                <button class="btn btn-primary" id="search-btn"> Search City</button>
-                <div id="error-msg" class="text-white"></div>
-            </div>
-            <div id="error-msg" class="text-white"></div>
-        </div>
-
-
-    </div>
-    </div>
-
-    <!-- <script src="js/jquery-3.3.1.slim.min.js"></script> -->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-    <script src="js/popper.min.js"></script>
-    <script src="js/bootstrap.js"></script>
-    <script src="js/weather.js">
-    </script>
-
-</body>
-
-</html>
