+++ client/public/index.html
@@ -45,19 +45,15 @@
             </div>
 
             <div class="col ">
-                <select type="text" id="search-area" placeholder="Type your city here" class="mt-input">
-                <option value="">Lagos, Nigeria</option>    
-                <option value="">Paris</option>    
-                <option value="">Washington</option>    
-                <option value="">London, uk</option>    
-                </select>
-                <button class="btn btn-primary" id="search-btn"> Search</button>
-
+                <input type="text" id="search-area" placeholder="Type your city here" class="mt-input" required="">
+                <button class="btn btn-primary" id="search-btn"> Search City</button>
                 <div id="error-msg" class="text-white"></div>
             </div>
-
-
+            <div id="error-msg" class="text-white"></div>
         </div>
+
+ 
+    </div>
     </div>
 
     <!-- <script src="js/jquery-3.3.1.slim.min.js"></script> -->
